# backend/src/main/resources/openapi.yaml
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/allergy":{"get":{"tags":["allergy-controller"],"operationId":"getAllergy","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Allergy"}}}}}}},"put":{"tags":["allergy-controller"],"operationId":"updateAllergyList","responses":{"200":{"description":"OK"}}},"post":{"tags":["allergy-controller"],"operationId":"assignAllergy","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Allergy"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"delete":{"tags":["allergy-controller"],"operationId":"removeAllergy","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Allergy"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/allergen":{"get":{"tags":["allergen-controller"],"operationId":"getAllergen","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Allergen"}}}}}},"put":{"tags":["allergen-controller"],"operationId":"updateAllergen","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Allergen"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["allergen-controller"],"operationId":"createAllergen","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Allergen"}}}}}},"delete":{"tags":["allergen-controller"],"operationId":"deleteAllergen","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/user/updatePassword":{"post":{"tags":["user-controller"],"operationId":"changeUserPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordChangeContainer"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/user/savePassword":{"post":{"tags":["user-controller"],"operationId":"savePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordResetContainer"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/user/resetPassword":{"post":{"tags":["user-controller"],"operationId":"resetPassword","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/update-user":{"post":{"tags":["user-controller"],"operationId":"updateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserMeta"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserMeta"}}}}}}},"/signup/{password}":{"post":{"tags":["user-controller"],"operationId":"signUp","parameters":[{"name":"password","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPrincipal"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/create-allergy":{"post":{"tags":["allergy-controller"],"operationId":"createAllergy","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Allergy"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Allergy"}}}}}}},"/assign-allergens":{"post":{"tags":["allergy-controller"],"operationId":"assignAllergens","parameters":[{"name":"allergyId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Allergen"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/assign-allergen":{"post":{"tags":["allergy-controller"],"operationId":"assignAllergen","parameters":[{"name":"allergyId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"allergenId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/user/changePassword":{"get":{"tags":["user-controller"],"operationId":"showChangePasswordPage","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/restaurantRating":{"get":{"tags":["restaurant-rating-controller"],"operationId":"getRestaurantRating","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RestaurantRating"}}}}}}}},"/current-user":{"get":{"tags":["user-controller"],"operationId":"getUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserContainer"}}}}}}},"/allergens":{"get":{"tags":["allergen-controller"],"operationId":"getAllergens","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Allergen"}}}}}}}},"/allergen-by-allergy":{"get":{"tags":["allergen-controller"],"operationId":"getAllergenByAllergy","parameters":[{"name":"allergyId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Allergen"}}}}}}}},"/allAllergies":{"get":{"tags":["allergy-controller"],"operationId":"getAllAllergies","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Allergy"}}}}}}}},"/update-allergy":{"delete":{"tags":["allergy-controller"],"operationId":"updateAllergy","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Allergy"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/delete-allergy":{"delete":{"tags":["allergy-controller"],"operationId":"deleteAllergy","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Allergy"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Allergen":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"allergy":{"$ref":"#/components/schemas/Allergy"}}},"Allergy":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"allergens":{"type":"array","items":{"$ref":"#/components/schemas/Allergen"}}}},"PasswordChangeContainer":{"type":"object","properties":{"oldPassword":{"type":"string"},"newPassword":{"type":"string"}}},"PasswordResetContainer":{"type":"object","properties":{"token":{"type":"string"},"newPassword":{"type":"string"}}},"UserMeta":{"type":"object","allOf":[{"$ref":"#/components/schemas/UserMetaBase"},{"type":"object","properties":{"allergies":{"type":"array","items":{"$ref":"#/components/schemas/Allergy"}}}}]},"Address":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"street":{"type":"string"},"city":{"type":"string"},"zip":{"type":"string"}}},"Authority":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"authority":{"type":"string"}}},"Menu":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"Restaurant":{"type":"object","allOf":[{"$ref":"#/components/schemas/UserMetaBase"},{"type":"object","properties":{"address":{"$ref":"#/components/schemas/Address"},"description":{"type":"string"},"menu":{"$ref":"#/components/schemas/Menu"}}}]},"UserMetaBase":{"required":["type"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"type":{"type":"string"}},"discriminator":{"propertyName":"type"}},"UserPrincipal":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"userMeta":{"oneOf":[{"$ref":"#/components/schemas/Restaurant"},{"$ref":"#/components/schemas/UserMeta"}]},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/Authority"}},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"},"enabled":{"type":"boolean"}}},"RestaurantRating":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"rating":{"type":"integer","format":"int32"},"comment":{"type":"string"},"user":{"$ref":"#/components/schemas/UserPrincipal"},"restaurant":{"$ref":"#/components/schemas/Restaurant"}}},"UserContainer":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserMeta"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/Authority"}}}}}}}